rewrite using pathlib
The rewrite using pathlib was done to help simplify the code, and use the
replacement for os.path for python 3.

Also during the rewrite a decision was made to try to replicate the behavior
of GNU stow behavior as much as possible.  The route the script was originally
going was going to be slightly different that stow but after some reflection
the way stow works is pretty dang elegant and trying to improve on it or
change how it works would be a little silly since so much though has been put
into how it works.