link: fix link source arg so it takes a string not a list
The link function was required to take a list with a single string to work as
expected. Otherwise passing a single string would cause strange behavior. This
makes the function work as expected by accepting just a single string four the
source argument.

Also fix the tests that hid the strange and unexpected behavior.