stow: move functions to correct abstraction level
Move validate_input() and

Move get_directory_contents() from AbstractBaseCommand() to AbstractBaseStow()
since it is not used in any capacity by the link command and is specialized
to stow and unstow.

simliarly replace validate_input() with a stub in AbstractBaseCommand() and
move it's implementation to AbstractBaseStow() since Link() implements it's
own validate_input() function.