unstow: fix issue where files could be deleted during folding
There was a issue where existing real files were not considered when deciding
weather to fold a directory (remove it and symlink an equivalent directory) so
folding was done and unconsidered files inside the directory were deleted.

So real files are now considered and a safer method of directory deletion is
done. Instead using shutil.rmtree instead use unstow the remaining links in the
directory use pathlib.Path.rmdir() so we only remove a directory if it's empty.