unstow: remove unneeded check if a dest is a symlink
Already unlinked messages were not getting generated because of a check if a
dest was a symlink in the case were both a source and dest in the unlink process
were both directories. This didn't seem to make sense so it was removed. It was
likely left over from a copy and paste of the stow command logic, that does
require the check.