unstow:fold: correctly delete empty directories
When unstowing all stowed packages directories created during the stowing
process via unfolding would not be deleted. This fix accounts for the case were
all the stowed packages will be unstowed, and cleans up the empty directories.

Also the unlink_actions_targets_parents variable was changed from a list to a
set to remove the duplicates inside it removing the need to check for
RemoveDirectory() actions.