stow/unstow: Check for errors before collecting actions
Check for additional errors while collecting actions, similar to how the initial
checks of the input to either a stow or unstow command are checked, but instead
of once, the check if performed every time actions are collected for stowing.

Also new consideration was taken for sources and dests dirs that have invalid
execute permissions.